UltraFeedback DataLoader Update
The ultrafeedback dataloader should run fine now. I also updated the testing dataloader file so you can test ultrafeedback with it too and tell it in the terminal which you want to test. I didn't change all that much in the code so hopefully it doesn't mess any of the github up